+++ NEWS.md
@@ -1,6 +1,6 @@
 # Unreleased
 - `propop::propop()` can now perform multiple projection scenarios 
-in a single run
+in a single run. 
 - The function `prepare_evaluation` now also runs with five age groups (0-19,
 20-39, 40-59, 60-79, 80 years and older). The output that is created by this 
 function can be used by the functions `compute_measures` and `aggregate_measures`